Update README
This project appears to have been finished - styles that the class is supposed to do are all in place.
While running the class yesterday without a monitor or projector, we accidentally steered the class to download
this zip as a starter kit.  

I've added a note that indicates that this is probably not the right place to start.